+++ lib/yearbook_web/components/layouts/backoffice.html.heex
@@ -1,5 +1,5 @@
 <div class="flex min-h-screen">
-  <YearbookWeb.BackofficeComponents.sidebar
+  <YearbookWeb.Sidebar.sidebar
     pages={YearbookWeb.Config.backoffice_pages()}
     current_page={Map.get(assigns, :current_page)}
   />
