--- lib/yearbook_web/live/backoffice/approvals_live/index.html.heex
@@ -1,1 +1,0 @@
-<.page title="Approvals"></.page>
