--- lib/yearbook_web/live/backoffice/year_live/index.ex
@@ -1,4 +1,0 @@
-defmodule YearbookWeb.YearLive do
-  use YearbookWeb, :live_view
-  def render(assigns), do: ~H"<div>Mock page Year</div>"
-end
